requirements.txt
# Core
python-dotenv
crewai

# Jupyter
notebook
ipykernel

# AI & LLM utilities
openai

# Data processing
pandas

# (Optional) Visualization
matplotlib
